api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "MyWorkout Api"
authors = ["Mateusz Marzec"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2.9"
gunicorn = "^20.1.0"
uvicorn = "^0.15.0"
sentry-sdk = "^1.4.3"
psycopg2-binary = "^2.9.2"
dj-database-url = "^0.5.0"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
django-filter = "^21.1"
django-allauth = "^0.46.0"
dj-rest-auth = "^2.1.11"
django-cors-headers = "^3.10.1"
django-autoslug = "^1.9.8"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.10b0"
isort = "^5.10.1"
autoflake = "^1.4"
mypy = "^0.910"
django-stubs = "^1.9.0"
djangorestframework-stubs = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | venv
  )/
'''

[tool.isort]
profile = "black"
skip = ["migrations", "venv"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["venv"]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = ["dj_database_url", "rest_framework_simplejwt.serializers", "rest_framework_simplejwt.views"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "api.settings.local"
ignore_missing_model_attributes = true
